delete test file only for py2

+++ recipes/effect/build.sh
@@ -1,2 +1,4 @@
-rm effect/_test_do_py3.py
+if [[ $PY_VER == 2* ]]; then
+    rm effect/_test_do_py3.py
+fi
 python setup.py install --single-version-externally-managed --record record.txt
